Story migration es v6

See merge request makeorg/platform/core-api!480

+++ core/src/main/scala/org/make/core/idea/indexed/Idea.scala
@@ -76,7 +76,7 @@ object IndexedIdea extends CirceFormatters {
   }
 }
 
-final case class IdeaSearchResult(total: Int, results: Seq[IndexedIdea])
+final case class IdeaSearchResult(total: Long, results: Seq[IndexedIdea])
 
 object IdeaSearchResult {
   implicit val encoder: ObjectEncoder[IdeaSearchResult] = deriveEncoder[IdeaSearchResult]
